refactor(OperandEditor): remove OperandOption folder
