--- packages/app-builder/src/components/Scenario/AstBuilder/AstBuilderNode/Operand/OperandEditor/OperandOption/index.ts
@@ -1,2 +1,0 @@
-export * from './ConstantOption';
-export * from './OperandOption';
